Move theme select and add layer to settings and remaining options except text
