+++ app/components/layer-settings/template.hbs
@@ -1,18 +1,86 @@
 <div class="p-md">
     <div class="layer-settings-inner">
-        <div class="layer-order">
+        <div class="layer-order inline-block">
             {{#if (gt index 0) }}
-                <button class="move-button" style="color:{{theme.color}};" {{action 'moveBefore' index}}> {{fa-icon 'long-arrow-up'}} Move up</button>
+                <button class="settings-button" style="color:{{theme.color}};" {{action 'moveBefore' index}}> {{fa-icon 'long-arrow-up'}} Move up</button>
             {{else}}
-                <button class="move-button" style="color:{{theme.color}};"  disabled> {{fa-icon 'long-arrow-up'}} Move up</button>
+                <button class="settings-button" style="color:{{theme.color}};"  disabled> {{fa-icon 'long-arrow-up'}} Move up</button>
             {{/if}}
 
             {{#if (lt index lastIndex)}}
-                <button class="move-button" style="color:{{theme.color}};" {{action 'moveAfter' index}}>{{fa-icon 'long-arrow-down'}} Move down</button>
+                <button class="settings-button" style="color:{{theme.color}};" {{action 'moveAfter' index}}>{{fa-icon 'long-arrow-down'}} Move down</button>
             {{else}}
-                <button class="move-button" style="color:{{theme.color}};" disabled>{{fa-icon 'long-arrow-down'}} Move down</button>
+                <button class="settings-button" style="color:{{theme.color}};" disabled>{{fa-icon 'long-arrow-down'}} Move down</button>
             {{/if}}
         </div>
-        {{theme-select themes=themes layer=layer}}
+
+
+        {{#each layer.settings.form as |item|}}
+            {{#if (eq item.type 'image')}}
+                <div class="inline-block">
+                    {{#bs-button}}
+                        <button class="settings-button">{{fa-icon 'image'}}</button>
+                        {{#bs-popover placement="bottom"}}
+                            {{input class="form-control" value=(get layer.settings.values item.value)  type="text"}}
+                        {{/bs-popover}}
+                    {{/bs-button}}
+                </div>
+
+
+            {{/if}}
+
+
+            {{#if (eq item.type 'increment')}}
+                <div class="inline-block">
+                    {{item.label}}
+                    <button class="btn btn-sm btn-default text-small" {{action 'changeSize' 'smaller' item}}>{{fa-icon 'fa-minus'}}</button>
+                    {{layer.settings.values.h1Size}}
+                    <button class="btn btn-sm btn-default text-big" {{action 'changeSize' 'bigger' item}}>{{fa-icon 'fa-plus'}}</button>
+                </div>
+            {{/if}}
+
+            {{#if (eq item.type 'alignment')}}
+                <div class="inline-block">
+                    {{item.label}}
+                    <div class="btn-group" role="group" aria-label="...">
+                        {{#each item.options as |option|}}
+                            <button type="button" class="btn btn-default {{if (eq layer.settings.values.alignment option) 'active'}}" {{action 'runOption' option item}} >{{option}}</button>
+                        {{/each}}
+                    </div>
+                </div>
+            {{/if}}
+
+        {{/each}}
+
+
+        <div class="theme-block inline-block">
+            Theme:
+            {{#each themes as |theme|}}
+                <div class="theme-box {{if (eq theme currentTheme) 'selected'}}"
+                     style="background:{{theme.background}}; color: {{theme.color}}"
+                    {{action 'updateTheme' theme}}
+                >A</div>
+            {{/each}}
+        </div>
+
+
+        <div class="inline-block" {{action 'showRemove'}}>
+            <button class="settings-button remove-layer-toggle">{{fa-icon "trash"}}</button>
+        </div>
+
     </div>
 </div>
+
+
+{{#if showRemoveModal}}
+    {{#bs-modal onSubmit=(action "removeLayer") onHidden=(action 'hideRemove') as |modal|}}
+        {{#modal.header}}
+            <h4 class="modal-title">Remove section?</h4>
+        {{/modal.header}}
+        {{#modal.body}}Are you sure you want to delete this section?{{/modal.body}}
+        {{#modal.footer as |footer|}}
+            {{#bs-button onClick=(action modal.close) type="default"}}Cancel {{/bs-button}}
+            {{#bs-button onClick=(action modal.submit) type="danger"}}Delete{{/bs-button}}
+        {{/modal.footer}}
+    {{/bs-modal}}
+{{/if}}
