--- app/components/theme-select/style.scss
@@ -1,12 +1,0 @@
-.theme-box {
-    width: 24px;
-    text-align: center;
-    border: 1px solid #eee;
-    cursor: pointer;
-    display: inline-block;
-}
-
-
-.theme-box:hover, .theme-box.selected {
-    border-color:#009688;
-}
