--- app/components/theme-select/template.hbs
@@ -1,11 +1,0 @@
-<div class="form-group clearfix">
-    <div class="theme-block">
-        Theme:
-        {{#each themes as |theme|}}
-            <div class="theme-box {{if (eq theme currentTheme) 'selected'}}"
-                 style="background:{{theme.background}}; color: {{theme.color}}"
-                 {{action 'updateTheme' theme}}
-            >A</div>
-        {{/each}}
-    </div>
-</div>
