+++ public/themes/theme_1.json
@@ -18,39 +18,9 @@
          },
          "form":[
             {
-               "type":"text",
+               "type":"image",
                "label":"Background image",
                "value":"backgroundImage",
-               "validation":null
-            },
-            {
-               "type":"text",
-               "label":"Lead text",
-               "value":"lead",
-               "validation":null
-            },
-            {
-               "type":"checkbox",
-               "label":"Show navigation",
-               "value":"showNavigation",
-               "validation":null
-            },
-            {
-               "type":"checkbox",
-               "label":"Fit image to section size",
-               "value":"backgroundCover",
-               "validation":null
-            },
-            {
-               "type":"checkbox",
-               "label":"Show title",
-               "value":"showTitle",
-               "validation":null
-            },
-            {
-               "type":"checkbox",
-               "label":"Show lead text",
-               "value":"showLead",
                "validation":null
             },
             {
@@ -62,7 +32,7 @@
                "validation":null
             },
             {
-               "type":"select",
+               "type":"alignment",
                "label":"Alignment",
                "value":"alignment",
                "options":[
