--- tests/integration/components/remove-layer/component-test.js
@@ -1,25 +1,0 @@
-import { moduleForComponent, test } from 'ember-qunit';
-import hbs from 'htmlbars-inline-precompile';
-
-moduleForComponent('remove-layer', 'Integration | Component | remove layer', {
-  integration: true
-});
-
-test('it renders', function(assert) {
-
-  // Set any properties with this.set('myProperty', 'value');
-  // Handle any actions with this.on('myAction', function(val) { ... });
-
-  this.render(hbs`{{remove-layer}}`);
-
-  assert.equal(this.$().text().trim(), '');
-
-  // Template block usage:
-  this.render(hbs`
-    {{#remove-layer}}
-      template block text
-    {{/remove-layer}}
-  `);
-
-  assert.equal(this.$().text().trim(), 'template block text');
-});
