--- tests/integration/components/theme-select/component-test.js
@@ -1,25 +1,0 @@
-import { moduleForComponent, test } from 'ember-qunit';
-import hbs from 'htmlbars-inline-precompile';
-
-moduleForComponent('theme-select', 'Integration | Component | theme select', {
-  integration: true
-});
-
-test('it renders', function(assert) {
-
-  // Set any properties with this.set('myProperty', 'value');
-  // Handle any actions with this.on('myAction', function(val) { ... });
-
-  this.render(hbs`{{theme-select}}`);
-
-  assert.equal(this.$().text().trim(), '');
-
-  // Template block usage:
-  this.render(hbs`
-    {{#theme-select}}
-      template block text
-    {{/theme-select}}
-  `);
-
-  assert.equal(this.$().text().trim(), 'template block text');
-});
